Move helper functions to command-utils.js
