+++ packages/core/lib/command-utils.js
@@ -6,6 +6,14 @@ const globalCommandOptions = require("./global-command-options");
 const debugModule = require("debug");
 const debug = debugModule("core:command:run");
 const commands = require("./commands/commands");
+const Config = require("@truffle/config");
+const Web3 = require("web3");
+const yargs = require("yargs");
+
+const defaultHost = "127.0.0.1";
+const managedGanacheDefaultPort = 9545;
+const managedGanacheDefaultNetworkId = 5777;
+const managedDashboardDefaultPort = 24012;
 
 // this function takes an object with an array of input strings, an options
 // object, and a boolean determining whether we allow inexact matches for
@@ -206,9 +214,90 @@ const displayGeneralHelp = () => {
     .showHelp();
 };
 
+const getConfiguredNetworkUrl = function (customConfig, isDashboardNetwork) {
+  const defaultPort = isDashboardNetwork
+    ? managedDashboardDefaultPort
+    : managedGanacheDefaultPort;
+  const configuredNetworkOptions = {
+    host: customConfig.host || defaultHost,
+    port: customConfig.port || defaultPort
+  };
+  const urlSuffix = isDashboardNetwork ? "/rpc" : "";
+  return `http://${configuredNetworkOptions.host}:${configuredNetworkOptions.port}${urlSuffix}`;
+};
+
+const deriveNetworkEnvironment = function (input) {
+  //detect config so we can get the provider and resolver without having to serialize
+  //and deserialize them
+  const { network, config, url } = yargs(input[0]).argv;
+  const detectedConfig = Config.detect({ network, config });
+
+  let configuredNetwork;
+
+  const configDefinesProvider =
+    detectedConfig.networks[network] &&
+    detectedConfig.networks[network].provider;
+
+  if (configDefinesProvider) {
+    // Use "provider" specified in the config to connect to the network
+    // along with the other specified network properties
+    configuredNetwork = {
+      ...detectedConfig.networks[network],
+      network_id: "*",
+      provider: detectedConfig.networks[network].provider
+    };
+  } else if (url) {
+    // Use "url" to configure network (implies not "develop" and not "dashboard")
+    configuredNetwork = {
+      network_id: "*",
+      url,
+      provider: function () {
+        return new Web3.providers.HttpProvider(url, {
+          keepAlive: false
+        });
+      }
+    };
+  } else {
+    // Otherwise derive network settings
+    const customConfig = detectedConfig.networks[network] || {};
+    const isDashboardNetwork = network === "dashboard";
+    const configuredNetworkUrl = getConfiguredNetworkUrl(
+      customConfig,
+      isDashboardNetwork
+    );
+    const defaultPort = isDashboardNetwork
+      ? managedDashboardDefaultPort
+      : managedGanacheDefaultPort;
+    const defaultNetworkId = isDashboardNetwork
+      ? "*"
+      : managedGanacheDefaultNetworkId;
+
+    configuredNetwork = {
+      // customConfig will spread only when it is defined and ignored when undefined
+      ...customConfig,
+      host: customConfig.host || defaultHost,
+      port: customConfig.port || defaultPort,
+      network_id: customConfig.network_id || defaultNetworkId,
+      provider: function () {
+        return new Web3.providers.HttpProvider(configuredNetworkUrl, {
+          keepAlive: false
+        });
+      }
+    };
+  }
+
+  detectedConfig.networks[network] = {
+    ...configuredNetwork
+  };
+
+  return detectedConfig;
+};
+
 module.exports = {
   displayGeneralHelp,
   getCommand,
   prepareOptions,
-  runCommand
+  runCommand,
+  getConfiguredNetworkUrl,
+  deriveNetworkEnvironment
 };
