+++ packages/core/lib/console-child.js
@@ -1,16 +1,8 @@
 const TruffleError = require("@truffle/error");
-const Config = require("@truffle/config");
-const Web3 = require("web3");
-const yargs = require("yargs");
 
 // we split off the part Truffle cares about and need to convert to an array
 const input = process.argv[2].split(" -- ");
 const inputStrings = input[1].split(" ");
-
-const defaultHost = "127.0.0.1";
-const managedGanacheDefaultPort = 9545;
-const managedGanacheDefaultNetworkId = 5777;
-const managedDashboardDefaultPort = 24012;
 
 // we need to make sure this function exists so ensjs doesn't complain as it requires
 // getRandomValues for some functionalities - webpack strips out the crypto lib
@@ -19,97 +11,35 @@ global.crypto = {
   getRandomValues: require("get-random-values")
 };
 
-//detect config so we can get the provider and resolver without having to serialize
-//and deserialize them
-const { network, config, url } = yargs(input[0]).argv;
-const detectedConfig = Config.detect({ network, config });
-
-function getConfiguredNetworkUrl(customConfig, isDashboardNetwork) {
-  const defaultPort = isDashboardNetwork
-    ? managedDashboardDefaultPort
-    : managedGanacheDefaultPort;
-  const configuredNetworkOptions = {
-    host: customConfig.host || defaultHost,
-    port: customConfig.port || defaultPort
-  };
-  const urlSuffix = isDashboardNetwork ? "/rpc" : "";
-  return `http://${configuredNetworkOptions.host}:${configuredNetworkOptions.port}${urlSuffix}`;
-}
-
-let configuredNetwork;
-
-const configDefinesProvider =
-  detectedConfig.networks[network] && detectedConfig.networks[network].provider;
-
-if (configDefinesProvider) {
-  // Use "provider" specified in the config to connect to the network
-  // along with the other specified network properties
-  configuredNetwork = {
-    ...detectedConfig.networks[network],
-    network_id: "*",
-    provider: detectedConfig.networks[network].provider
-  };
-} else if (url) {
-  // Use "url" to configure network (implies not "develop")
-  configuredNetwork = {
-    network_id: "*",
-    url,
-    provider: function () {
-      return new Web3.providers.HttpProvider(url, {
-        keepAlive: false
-      });
-    }
-  };
-} else {
-  // Otherwise derive network settings
-  const customConfig = detectedConfig.networks[network] || {};
-  const isDashboardNetwork = network === "dashboard";
-  const configuredNetworkUrl = getConfiguredNetworkUrl(
-    customConfig,
-    isDashboardNetwork
-  );
-  const defaultPort = isDashboardNetwork
-    ? managedDashboardDefaultPort
-    : managedGanacheDefaultPort;
-  const defaultNetworkId = isDashboardNetwork
-    ? "*"
-    : managedGanacheDefaultNetworkId;
-
-  configuredNetwork = {
-    // customConfig will spread only when it is defined and ignored when undefined
-    ...customConfig,
-    host: customConfig.host || defaultHost,
-    port: customConfig.port || defaultPort,
-    network_id: customConfig.network_id || defaultNetworkId,
-    provider: function () {
-      return new Web3.providers.HttpProvider(configuredNetworkUrl, {
-        keepAlive: false
-      });
-    }
-  };
-}
-
-detectedConfig.networks[network] = {
-  ...configuredNetwork
-};
-
-const { getCommand, prepareOptions, runCommand } = require("./command-utils");
-const command = getCommand({ inputStrings, options: {}, noAliases: false });
-const options = prepareOptions({
-  command,
-  inputStrings,
-  options: detectedConfig
-});
-
-runCommand(command, options)
-  .then(() => process.exit(0))
-  .catch(error => {
-    // Perform error handling ourselves.
-    if (error instanceof TruffleError) {
-      console.log(error.message);
-    } else {
-      // Bubble up all other unexpected errors.
-      console.log(error.stack || error.toString());
-    }
-    process.exit(1);
+function main() {
+  const {
+    getCommand,
+    prepareOptions,
+    runCommand,
+    deriveNetworkEnvironment
+  } = require("./command-utils");
+  const configuredNetwork = deriveNetworkEnvironment(input);
+  const command = getCommand({ inputStrings, options: {}, noAliases: false });
+  const options = prepareOptions({
+    command,
+    inputStrings,
+    options: configuredNetwork
   });
+
+  runCommand(command, options)
+    .then(() => process.exit(0))
+    .catch(error => {
+      // Perform error handling ourselves.
+      if (error instanceof TruffleError) {
+        console.log(error.message);
+      } else {
+        // Bubble up all other unexpected errors.
+        console.log(error.stack || error.toString());
+      }
+      process.exit(1);
+    });
+}
+
+main();
+
+module.exports = { main };
